requirements.txt
python-chess==1.999
gymnasium==0.29.1
torch>=2.0
torchvision>=0.15
onnx>=1.14
onnxruntime>=1.15
tqdm>=4.65
numpy>=1.24,<2
